Extracted create_person from app
